refactor planner into multiple pages
also includes some minor UI adjustments and some feature work to improve the date selector
