+++ frontend/.nuxtignore
@@ -1,0 +1,1 @@
+pages/**/*.ts
