+++ frontend/pages/group/mealplan/planner/types.ts
@@ -1,0 +1,6 @@
+import { ReadPlanEntry } from "~/lib/api/types/meal-plan";
+
+export type MealsByDate = {
+  date: Date;
+  meals: ReadPlanEntry[]
+}
